[Fix] Multiple Changes on Job_OfferJpaEntity
Changes were made in the IJob_Offer repository by placing a class that is searched by id of employers. In addition, in the Job_OfferController the aforementioned class was placed to return the jobs by employer id that will then be shared to be consumed. Finally, in Job_OfferJpaEntity it was changed from employer_id to employerId.

ID E.12